feat: shared overflow recovery + remove force-wipe
- Add overflow recovery to ChatSession base class:
  - _is_context_overflow_error() — abstract, providers override
  - _trim_context_one_round() — drops ~10% oldest non-system entries
  - _run_with_overflow_recovery() — retry loop with trim
  - _inject_overflow_notice() — post-recovery user message

- Anthropic: implement _is_context_overflow_error(), update send()
  and send_stream() to use shared recovery

- OpenAI: remove duplicate methods, use base class inheritance

- Remove force-wipe from turn.py:
  - No more context_forget() on pressure
  - Agent gets warnings and can molt voluntarily
  - If agent ignores warnings, API call fails and overflow recovery trims

DeepSeek, OpenRouter inherit from OpenAI.
MiniMax inherits from Anthropic.
Custom adapter delegates to the above.
